Feat/pool fee in api (#39)
* Initial poolFee commit

This is defined by the following in the pool_config

```
    "poolFee": 0.2,
    "_comment_poolFee": "This is only used for API purpose to display pool fee percentage",
```

* Added poolFee function

This function looks at the user defined `poolFee` in their pool_config and if it doesn't exist it will calculate it based on `rewardRecipients`

* Fix to show %

* Added poolFee

* Added poolFee

* Added poolFee

* Added poolFee

* Added poolFee

* Added poolFee

* Added poolFee

* Added poolFee

* Added poolFee

* Added poolFee

* Added poolFee

* Spacing fix

* Changing gears -- instead of providing additional configuration, we're
just going to expose the rewardRecipients. This allows the consuming
client to filter community donation addresses, etc. as needed.

In addition, I've replaced all tabs in libs/stats.js with spaces.