Add payment success check
Due to the nature of blockchains, some blocks get orphaned. We see this
on all our pools and the pool is coded to account for orphaned blocks
when it comes to calculating payouts. Unfortunately, the pool doesn't
check to ensure payouts made didn't get caught up in an orphaned block.
This commit adds a final step at the end of the payment loop that
checks the most recent 25th, 50th and 75th payment to ensure it
succeeded (confirmations != -1).

In the case that the daemon returns -1 for confirmations, the payment is
rebroadcast and the database is updated with the new payment.

I am currently running this code on my pool. I haven't tested it with
s-nomp and it would be a bit of a pain to test. If anyone is interested
in testing it by modifying the code to assume a failure from the daemon,
please do! I'd like at the very least to get a test on payment
processing to ensure it doesn't throw any errors.